Password Forget Page and Change Form Refactor
Password Forget Page and Change Form Refactor
Plus small fixes
