+++ frontend/src/components/FormComponents/index.js
@@ -3,13 +3,15 @@ import ContactInfoForm from "./ContactInfoForm";
 import MemberInfoForm from "./MemberInfoForm";
 import MemberTypeForm from "./MemberTypeForm";
 import CheckoutForm from "./CheckoutForm";
+import PasswordChangeForm from "./PasswordChangeForm";
 
 export default ContactInfoForm;
 
 export {
   AddressForm,
+  CheckoutForm,
   ContactInfoForm,
   MemberInfoForm,
   MemberTypeForm,
-  CheckoutForm
+  PasswordChangeForm,
 }
